[FLINK-25282][table-planner][table-runtime] Moved some functions required at runtime from table-planner to table-runtime
