+++ flink-table/flink-table-runtime/src/main/java/org/apache/flink/table/runtime/operators/misc/FilterAllFlatMapFunction.java
@@ -1,0 +1,48 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.table.runtime.operators.misc;
+
+import org.apache.flink.annotation.Internal;
+import org.apache.flink.api.common.functions.FlatMapFunction;
+import org.apache.flink.api.common.typeinfo.TypeInformation;
+import org.apache.flink.api.java.typeutils.ResultTypeQueryable;
+import org.apache.flink.table.data.RowData;
+import org.apache.flink.table.runtime.typeutils.InternalTypeInfo;
+import org.apache.flink.util.Collector;
+
+/** Function filtering out all the input records. */
+@Internal
+public class FilterAllFlatMapFunction
+        implements FlatMapFunction<RowData, RowData>, ResultTypeQueryable<RowData> {
+    private static final long serialVersionUID = 1L;
+
+    private final InternalTypeInfo<RowData> outputTypeInfo;
+
+    public FilterAllFlatMapFunction(InternalTypeInfo<RowData> inputTypeInfo) {
+        this.outputTypeInfo = inputTypeInfo;
+    }
+
+    @Override
+    public void flatMap(RowData value, Collector<RowData> out) {}
+
+    @Override
+    public TypeInformation<RowData> getProducedType() {
+        return outputTypeInfo;
+    }
+}
